+++ NEWS.md
@@ -1,5 +1,17 @@
-# teal 0.10.1.9001
+# teal 0.10.1.9006
 
+### Bug fixes
+* `get_call()` function of `CallableFunction` now returns call with namespace included.
+* Deprecated `root_modules` function, users should use `modules` directly inside `init`.
+
+### Breaking changes
+* Due to deprecation of `root_modules` any `label` argument to `modules` must be explicitly named. For example `modules("lab", mod1, mod2)` should be replaced with `modules(label = "lab", mod1, mod2)`.
+
+### Miscellaneous
+* New argument `ordered` in the `select_spec()` to flag whether order of the selection should be tracked.
+* Minor changes to internals of `teal`: main module panel now has fixed shiny name `root` and the active tab is named `active_tab` not `Active_tab`. 
+
+# teal 0.10.1
 ### Breaking changes
 * Changed the `HTML` identifiers of `teal` modules - now each nested module receives its own `shiny` namespace.
 * Deprecated all functions related to connectors that have been moved to their own separate packages.